docs and refactor
